backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
python-dotenv==1.0.1
langgraph>=0.2.0
langchain-core>=0.3.0
google-generativeai>=0.8.0
openai>=1.0.0
